# docs/openapi.courtly.yaml
openapi: 3.1.0
info:
  title: Courtly API
  version: 1.0.0
  description: >
    Full-scope API for Courtly tennis court booking platform.
    Includes auth, courts, day/week availability, hold-based booking,
    user cabinet, moderator/admin dashboard, dynamic RBAC/policy management,
    and event log operations.
servers:
  - url: https://api.courtly.local
    description: Production
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and issue tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid credentials
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "401":
          description: Invalid refresh token
  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      security: []
      description: >
        Sends password reset link if account exists.
        Always returns generic accepted response to prevent user enumeration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "202":
          description: Accepted (generic response)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "If an account exists, password reset instructions were sent."
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Confirm password reset using one-time token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "204":
          description: Password reset successfully
        "400":
          description: Invalid or expired token
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Change current password
      responses:
        "204":
          description: Password changed
        "400":
          description: Invalid input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
...
  /api/admin/roles:
    get:
      tags: [RBAC]
      summary: List roles
      x-rbac:
        permission: roles:read
        ownership: admin_any
      responses:
        "200":
          description: Roles
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoleResponse"
    post:
      tags: [RBAC]
      summary: Create role
      x-rbac:
        permission: roles:create
        ownership: admin_any
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleResponse"
  /api/admin/policies:
    post:
      tags: [RBAC]
      summary: Create policy rule
      x-rbac:
        permission: policies:create
        ownership: admin_any
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    ChangePasswordRequest:
      type: object
      required: [old_password, new_password]
      properties:
        old_password:
          type: string
        new_password:
          type: string
          minLength: 8
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [reset_token, new_password]
      properties:
        reset_token:
          type: string
          description: One-time password reset token from email link.
        new_password:
          type: string
          minLength: 8
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: bearer
    AccessTokenResponse:
      type: object
      required: [access_token, token_type]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
...
    RoleCreateRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
    RoleResponse:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
    PolicyCreateRequest:
      type: object
      required: [name, effect, condition]
      properties:
        name: { type: string }
        effect:
          type: string
          enum: [allow, deny]
        condition:
          type: string
          description: "Policy expression evaluated by policy engine."
    PolicyResponse:
      type: object
      required: [id, name, effect, condition]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        effect: { type: string, enum: [allow, deny] }
        condition: { type: string }
...